Move content loading to activity
