+++ app/src/main/java/org/jellyfin/androidtv/playback/nextup/UpNextItemData.kt
@@ -1,9 +1,11 @@
 package org.jellyfin.androidtv.playback.nextup
 
+import android.graphics.Bitmap
+
 data class UpNextItemData(
 	val id: String,
 	val title: String,
 	val description: String?,
-	val backdrop: String?,
-	val thumbnail: String?
+	val backdrop: Bitmap?,
+	val thumbnail: Bitmap?
 )
